+++ parsers/pstring/string.go
