+++ parsers/pstring/string_test.go
